requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.19
boto3==1.35.87
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.0
bcrypt==4.2.1
